fix: reconcile PlayerHealthComponent API with consumer code
- Add legacy delegate shims (OnTakeDamage, OnPlayerDeath) with int32 signatures
- Add backward-compat method shims (GetTotalHitsTaken, GetMaxHealth, ResetHealth, etc.)
- Update delegate handler signatures in RunnerCharacter and HealthBarWidget to float
- Fix DEFINE_LOG_CATEGORY typo in PlayerHealthComponent.cpp
- Replace nonexistent World::GetPawnsOfType with TActorIterator in SideRunnerGameMode
- Fix %d/%.1f format mismatch in HealthBarWidget log

+++ Source/SideRunner/HealthBarWidget.cpp
@@ -56,14 +56,14 @@ void UHealthBarWidget::NativeTick(const FGeometry& MyGeometry, float InDeltaTime
 // Health Component Event Handlers
 // ============================================================================
 
-void UHealthBarWidget::OnHealthChanged(int32 NewHealth, int32 NewMaxHealth)
+void UHealthBarWidget::OnHealthChanged(float NewHealth, float NewMaxHealth)
 {
 	// Update cached values
-	CurrentHealth = static_cast<float>(NewHealth);
+	CurrentHealth = NewHealth;
 	MaxHealth = static_cast<float>(NewMaxHealth);
 
 #if UE_BUILD_DEVELOPMENT
-	UE_LOG(LogSideRunner, VeryVerbose, TEXT("HealthBarWidget: Health changed to %d / %d (%.1f%%)"),
+	UE_LOG(LogSideRunner, VeryVerbose, TEXT("HealthBarWidget: Health changed to %.1f / %.1f (%.1f%%)"),
 		NewHealth, NewMaxHealth, GetHealthPercent() * 100.0f);
 #endif
